pyproject.toml
[project]
name = "my-pos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "niquests>=3.15.2",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]
